+++ semver-project-gradle-plugin/testFixtures/kotlin/com/javiersc/semver/project/gradle/plugin/AssertVersion.kt
@@ -1,6 +1,7 @@
 package com.javiersc.semver.project.gradle.plugin
 
 import com.javiersc.gradle.version.GradleVersion
+import com.javiersc.gradle.version.GradleVersion.CheckMode.Significant
 import com.javiersc.gradle.version.GradleVersionException
 import com.javiersc.kotlin.stdlib.AnsiColor.Foreground.BrightYellow
 import com.javiersc.kotlin.stdlib.AnsiColor.Foreground.Purple
@@ -58,13 +59,13 @@ internal fun File.assertVersion(
             buildVersion.startsWith(version).shouldBeTrue()
             buildTagVersion.startsWith("$prefix$version").shouldBeTrue()
             buildVersion.shouldContain("+")
-            shouldThrow<GradleVersionException> { GradleVersion(buildVersion) }
+            shouldThrow<GradleVersionException> { GradleVersion(buildVersion, Significant) }
         }
         Insignificant.Dirty -> {
             buildVersion.startsWith(version).shouldBeTrue()
             buildTagVersion.startsWith("$prefix$version").shouldBeTrue()
             buildVersion.shouldContain("+").shouldContain("DIRTY")
-            shouldThrow<GradleVersionException> { GradleVersion(buildVersion) }
+            shouldThrow<GradleVersionException> { GradleVersion(buildVersion, Significant) }
         }
         else -> {
             buildVersionFile